feature/interview
During this assessment, you will be given a set of data tables and tasks that will test your technical proficiency in programming tools, database design, and problem-solving. We encourage you to approach each challenge with a thorough understanding of the problem at hand and to use your creativity and critical thinking skills to come up with effective solutions.